@@ -7,7 +7,7 @@
     <Nullable>enable</Nullable>
     <PathMap>$([System.IO.Path]::GetFullPath("$(MSBuildThisFileDirectory)"))=./</PathMap>
     <RootNamespace>Hirameku.$(MSBuildProjectName.Replace(" ", "_"))</RootNamespace>
-    <TargetFramework>net6.0</TargetFramework>
+    <TargetFramework>net8.0</TargetFramework>
     <TreatWarningsAsErrors>True</TreatWarningsAsErrors>
   </PropertyGroup>
 </Project>
