@@ -161,7 +161,7 @@ public class CardValidatorTests
         };
     }
 
-    private static IEnumerable<Meaning> GetFakeMeanings(int length = 1)
+    private static List<Meaning> GetFakeMeanings(int length = 1)
     {
         var meaningFaker = new Faker<Meaning>()
             .RuleFor(m => m.Example, string.Empty)
